+++ src/Integerise.h
@@ -5,4 +5,4 @@
 #include <cstdlib>
 
 // Given pop and real number densities (sum = 1), produce integer frequencies with minimal mean squuare error
-std::vector<int> integeriseMarginalDistribution(const std::vector<double>& p, size_t pop, double& mse);
+std::vector<int> integeriseMarginalDistribution(const std::vector<double>& p, int pop, double& mse);
